fix: useSearchParams 관련 Suspense 오류 해결을 위한 구조 변경
